MapSetup
Bomberman:  calling g_map to render it

index: changed canvas resolution for testing purposes, we can scale it up and down with a few tweaks.

plauyerjs:

Moved playermoevment to a function, changed constant movement to "step" can also be seen as characterspeed.

MapSetup:

added a map made out of an array:
when we have sprites for the tiles it shouldn't be difficult to replace rectangles with appriate sprites.

some other misc map stuff i was tinkering with that i'll clean up later. they are commented out cause they aren't being used.