Sound /  Map / UI
Made some changes to the map generation

Made some changes to general sound

It might be better to have the previous bomb sound because it is a single audio object that gets played on explosion.

But my implementaion (derived from what Emil did with sound) creates a new sound for every bomb explosion. meaning they will overlap however often g_sounds.bombExplosion.play() gets called.

Which seems to be like 5 times.

other than that I also made some configuration around the UI and it is now ready to be updated with whatever we want to display on the canvas.