Win condition // some map tweaks// sound tweaks and some UI tweaks
added addition text in UI, all very plain right now. We can add sprite of a bomb instead of "bomb" text and so on.

added more sounds and also made some sounds play when picking up things.

added an interaction function for the player. on pressing E it will check if any adjcent tiles are 'interactable'. Currently it is set to remove the gates.

Upon reaching collecting 4 keyes, "you wiN" is written in the middle of the screen.

added a gates javascript file, this is where questions are supposed to be asked and then displayed with the UI script.

Still have not fully implemented it. but that is the idea.